requirements.txt
Flask
Flask-Cors
gunicorn
google-generativeai
openai